requirements.txt
python-dotenv
pyTelegramBotAPI
flask
requests